Add files via upload
FRONT END:
HOMEPAGE:
When the user logs on to the website, they will be directed to the homepage,
from where they can further proceed
to know more about the travel
agency, the contact information or the
packages available by clicking on one
of the options on the navigation bar.
ABOUT US:
Upon clicking on the Nav item “About
Us”, the user views the navigation link
where they can read about the
enterprise, its services and legacy.
CONTACT US:
The Nav item “Contact Us” is
navigating the user to the link where
the contact information of the
franchise is available.
PACKAGES:
The Nav item “Packages” has a drop-down menu that
includes the packages that the travel franchise provides to
its customers. The user can choose to view the details of
any of the provided travel packages. In our case there are
four packages available
• Murree package (Couple package)
• Hunza package (Family package)
• Skardu Package(Family package)
• Naran Package(Family package)
The Hypertext reference to each of these
drop-down items is their respective preview
pages where the website displays the details
about the location, number of days the
customer gets to stay there, transport, hotel,
the ambience and the price per person for the
purchase.
BOOKING:
When the user clicks on “Book Now”, they will
be redirected to the hypertext reference of the
booking page where they can select their
desired package, desired vehicle for
transportation and the number of persons
travelling. Note that a couple package like
Murree can include only two persons, if the user
wishes to book a trip for more than two people,
they have to purchase a family package which has
to include more than two persons. Although
booking confirmation requires the user to be
logged in.
If the user is already logged in, he will be
redirected to check out on clicking “confirm”
where he can see the details of their booking
and the total amount to be paid. Checkout also contains a delete button which
when clicked by the customer deletes the customer’s current booking details
from the database.
LOGIN MODAL:
In order to proceed with the booking
confirmation, the website requires user login.
The login modal uses a form where it takes
the user’s email and password and then
checks the database for all existing customers
and on successfully finding the data of the
customer, it allows them to proceed with
their booking. On unsuccessful login attempt it echoes error message “Invalid
email or password”.
SIGN-UP MODAL:
A new customer can start off by signing
up, the sign-up modal includes a form
that asks the user for its Name, Email ID,
password, confirm password to compare,
contact number, alternate contact
number and address. Upon clicking on
“Create account these credentials are
added to the database if the user has not
already signed up with the franchise.
“Already a member” button redirects the
customer to the login modal.
CHANGE PASSWORD MODAL:
A customer can change his/her password by
selecting the change password option from the
dropdown menu on the home page, the option
redirects the customer to the change password
modal which then asks user to enter password
in old password, new password and confirm
new password fields and then updates the
password in the database.
BACK-END:
After normalization, we have 11 tables to implement in our database “tourism”.
We have used phpmyadmin to create our database. Given below are the tables
and their working.
1.CUSTOMER(customer_id, customer_name, customer_email,
customer_password,customer_addres:
This table contains id, name, email,
password and address of the customer.
These information are inserted into this
table whenever a new customer signs
in.
2.CUSTOMER_INFO(customer_id, customer_contact_no):
This table holds the record of the customer’s id and the customer’s contact
number. The need for a separate
table for the customer’s contact
number is that it is a multivalued
attribute which causes insertion
and updation anomalies in our
database.
3.PACKAGE(package_id, package_name, package_type, package_desc,
amount_per_person):
This table holds the record of the
package’s id
, name, type, description and the
amount per person for the specific
package. For now, our database
contains four different packages but we can also add or remove a package.
4. BOOKING(booking_id, booking_date, booking_type, booking_desc,
person, booking_total):
This table holds the record of the complete booking details of customers. It
contains booking ids with respective
package and customer ids,
here package and customer id
are the foreign keys. It also
contains details like booking
date, type, description, the
number of persons a customer has entered for a certain package and the
total amount of booking. For Murree package there is a limitation that it is
only for 2 persons as it is a couple package.
5. HOTEL(( hotel_id, hotel_name, hotel_type, hotel_desc):
This table contains hotel details like its id, name, type and
description.Whenever a customer
books a package a hotel is assigned to
the customer according to their
respective package.
6. HOTEL_INFO(( hotel_id, hotel_address):
This table holds the record of the hotel’s id
and the hotel’s address. Like the customer
contact number, hotel_address is also a
multivalued attribute, that’s why we have
made a separate table for it.
6.TRANSPORTATION(transportation_id,
transportation_type,transportation_desc):
This table contains transportation
details like its id, type and
description. We have limited our
database to only 3 vehicles and a
customer has to choose one out of
these three vehicles at the time of
booking.
The rest 4 tables are of relationship sets which are as follows:
1.CUSTOMER-PACKAGE(customer_id[FK],package_id[FK]):
As customer and package has a many-to-many relation so its relationship
set contains the primary keys of both the
tables and these primary keys behave like
the foreign keys in this relationship set
table.
2.CUSTOMER_BOOKING(customer_id[FK],booking_id[FK]:
Customer and booking also has a many-to-many relation, so its relationship
set contains the primary keys of both the
respective tables and these primary keys
behave like the foreign keys in this
relationship set table.
3.BOOKING_HOTEL(booking_id[FK],hotel_id[F]):
After some changes in the er diagram, Booking and hotel now has a binary
many-to-many relation so its relationship set
contains the primary keys of both the
respective tables and these primary keys
behave like the foreign keys in this relationship set table.
4.BOOKING_TRANSPORTATION(booking_id[FK],transportati
on_id[F]):
After some changes in the er diagram, booking and transportation now has
a binary many-to-many relation so its
relationship set contains the primary keys of
both the respective tables and these
primary keys behave like the foreign keys in this relationship set table.